requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
archspec @ file:///croot/archspec_1697725767277/work
boltons @ file:///work/ci_py311/boltons_1677685195580/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
build==1.1.1
CacheControl==0.14.0
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cleo==2.1.0
click==8.1.7
conda @ file:///croot/conda_1701719518285/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
contourpy==1.2.0
crashtest==0.4.1
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
diffusers==0.4.0
distlib==0.3.8
distro @ file:///croot/distro_1701455004953/work
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.1
idna @ file:///work/ci_py311/idna_1676822698822/work
importlib_metadata==7.0.2
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
jeepney==0.8.0
Jinja2==3.1.3
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
keyring==24.3.1
kiwisolver==1.4.5
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
menuinst @ file:///croot/menuinst_1702390294373/work
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
orjson==3.9.15
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.2.1
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.10.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==5.19.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycodestyle==2.11.1
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.2
pyproject_hooks==1.0.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
requests-toolbelt==1.0.0
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
safetensors==0.4.2
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
tomlkit==0.12.4
torch==2.2.1
tqdm @ file:///croot/tqdm_1679561862951/work
transformers==4.39.1
triton==2.2.0
trove-classifiers==2024.3.3
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
zipp==3.18.1
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
